Remove simplejson as a dependency for python 2.6+
Python 2.6 introduced the ``json`` module into the core language and so ``simplejson`` is only required in python versions prior to python 2.6. This fix ensures that pip will only install simplejson if the version of python being run is less than 2.6. This stops the un-needed installation of a library that wont ever actually get used by authy if a newer version of python is running.